:recycle: refactor: Move openmp.c to openmp directory and update README with compilation instructions. Update at cuda algorithm
