+++ openmp/gabriel_almeida_rodrigues_pereira_entrega1.pdf
